Add a new port & note password length requirement
Valheim Dedicated Server now seems to require port 2458 to be open in
addition to the previously required ports, this has been added to the
container definition and noted in the README.

The server also seems to have a minimum password length of 5 characters.
Shorter passwords will result in seemingly unrelated NREs when trying
to run the server. This is now noted in the README.